fix(primtime) increased the message buffer
If a message is larger than the receive buffer,
and the receive mode is "line", gen_tcp will
receive another batch of data up to the buffer size.
It seems that by default the buffer size matches the
receive buffer size --> it is too small for packages
that are larger than around 1kb.